requirements.txt
pyzotero>=1.5.0
python-dotenv>=1.0.0
